Renaming IHS_toolkit by ihs_toolkit
